@@ -1,8 +1,7 @@
 import importlib
 import sys
 from importlib.util import LazyLoader, find_spec, module_from_spec
-from types import ModuleType
-from typing import Literal
+from typing import Literal, Optional, Any
 
 
 def reload(module: str) -> None:
@@ -22,7 +21,7 @@ def reload(module: str) -> None:
     return importlib.import_module(module)
 
 
-def smart_import(name: str, mode: Literal["lazy"] | None = None) -> ModuleType:
+def smart_import(name: str, mode: Literal["lazy"] | None = None, attribute: Optional[str] = None) -> Any:
     """Import a module with optional lazy loading.
 
     This function imports a module by name. If the module is already
@@ -35,9 +34,11 @@ def smart_import(name: str, mode: Literal["lazy"] | None = None) -> ModuleType:
         name (str): The name of the module to import.
         mode (Literal["lazy"] | None, optional): If "lazy", the module will
             be imported using a lazy loader. Defaults to None.
+        attribute (str, optional): The name of an attribute, class, or symbol
+            to import from the module. Defaults to None.
 
     Returns:
-        ModuleType: The imported module.
+        Any: The imported module or the specified attribute from the module.
 
     Raises:
         NameError: If the module cannot be found or imported.
@@ -54,10 +55,18 @@ def smart_import(name: str, mode: Literal["lazy"] | None = None) -> ModuleType:
         module = module_from_spec(spec)
         sys.modules[name] = module
         loader.exec_module(module)
-        return module
 
     try:
-        return importlib.import_module(name)
+        module = importlib.import_module(name)
     except ImportError as e:
         msg = f"Module {name} not found"
         raise NameError(msg) from e
+
+    if attribute:
+        try:
+            return getattr(module, attribute)
+        except AttributeError as e:
+            msg = f"Attribute `{attribute}` not found in module `{name}`"
+            raise NameError(msg) from e
+
+    return module
