@@ -56,15 +56,13 @@ def test_anthropic_backend_create_completion(anthropic_backend):
 
 def test_anthropic_backend_with_context(anthropic_backend):
     context = [Message("Hello"), Message("How are you?")]
-    response = anthropic_backend.predict(
-        Message("What's the weather like?"), context=context)
+    response = anthropic_backend.predict(Message("What's the weather like?"), context=context)
     assert response == mock_anthropic_response
 
 
 def test_anthropic_backend_with_image(anthropic_backend):
     image = Image(size=(224, 224))
-    response = anthropic_backend.predict(
-        Message(["Describe this image", image]), context=[])
+    response = anthropic_backend.predict(Message(["Describe this image", image]), context=[])
     assert response == mock_anthropic_response
 
 
@@ -84,11 +82,9 @@ def test_anthropic_backend_stream(anthropic_backend):
         return MockStream()
 
     anthropic_backend.client.messages.stream = mock_stream
-    stream_generator = anthropic_backend.stream(
-        Message("Stream this message"), context=[])
+    stream_generator = anthropic_backend.stream(Message("Stream this message"), context=[])
     results = list(stream_generator)
-    expected_results = ["Streaming response part 1",
-                        "Streaming response part 2", ""]
+    expected_results = ["Streaming response part 1", "Streaming response part 2", ""]
 
     assert results == expected_results
 
